+++ components/hal/esp32c6/include/hal/i2c_ll.h
@@ -598,7 +598,7 @@ static inline void i2c_ll_get_stop_timing(i2c_dev_t *hw, int *setup_time, int *h
 __attribute__((always_inline))
 static inline void i2c_ll_write_txfifo(i2c_dev_t *hw, const uint8_t *ptr, uint8_t len)
 {
-    for (int i = 0; i < len; i++) {
+    for (int i = 0; i < (int)len; i++) {
         HAL_FORCE_MODIFY_U32_REG_FIELD(hw->data, fifo_rdata, ptr[i]);
     }
 }
@@ -615,7 +615,7 @@ static inline void i2c_ll_write_txfifo(i2c_dev_t *hw, const uint8_t *ptr, uint8_
 __attribute__((always_inline))
 static inline void i2c_ll_read_rxfifo(i2c_dev_t *hw, uint8_t *ptr, uint8_t len)
 {
-    for (int i = 0; i < len; i++) {
+    for (int i = 0; i < (int)len; i++) {
         ptr[i] = HAL_FORCE_READ_U32_REG_FIELD(hw->data, fifo_rdata);
     }
 }
@@ -630,7 +630,7 @@ static inline void i2c_ll_read_rxfifo(i2c_dev_t *hw, uint8_t *ptr, uint8_t len)
  */
 static inline void i2c_ll_write_by_nonfifo(i2c_dev_t *hw, uint8_t ram_offset, const uint8_t *ptr, uint8_t len)
 {
-    for (int i = 0; i < len; i++) {
+    for (int i = 0; i < (int)len; i++) {
         hw->txfifo_mem[i + ram_offset] = ptr[i];
     }
 }
@@ -645,7 +645,7 @@ static inline void i2c_ll_write_by_nonfifo(i2c_dev_t *hw, uint8_t ram_offset, co
  */
 static inline void i2c_ll_read_by_nonfifo(i2c_dev_t *hw, uint8_t ram_offset, uint8_t *ptr, uint8_t len)
 {
-    for (int i = 0; i < len; i++) {
+    for (int i = 0; i < (int)len; i++) {
         ptr[i] = hw->rxfifo_mem[i + ram_offset];
     }
 }
